Refactor RTC manager implementations
