+++ src/rtc_manager.cpp
@@ -1,5 +1,9 @@
 #include "rtc_manager.h"
 
+#include "config.h"
+
+#include <time.h>
+
 void enableRTC() {
   digitalWrite(GPIO_RS485_ENABLE, HIGH);
   Serial.flush();
@@ -44,3 +48,19 @@ void setRTCFromWeb(String date, String time) {
     Serial.println(F("🕒 RTC wurde aktualisiert!"));
 }
 
+void syncTimeWithNTP() {
+    Serial.println(F("🔄 Synchronisiere Zeit mit NTP..."));
+    configTime(0, 0, "pool.ntp.org", "time.nist.gov");
+
+    struct tm timeinfo;
+    if (getLocalTime(&timeinfo, 10000)) {
+        enableRTC();
+        rtc.adjust(DateTime(timeinfo.tm_year + 1900, timeinfo.tm_mon + 1,
+                            timeinfo.tm_mday, timeinfo.tm_hour,
+                            timeinfo.tm_min, timeinfo.tm_sec));
+        enableRS485();
+        Serial.println(F("✅ NTP Synchronisierung erfolgreich!"));
+    } else {
+        Serial.println(F("❌ NTP Zeit konnte nicht abgerufen werden!"));
+    }
+}
